Wielded lights as a drop in replacement for illumination (#89)
* Replaced illumination mod with wielded_light mod

drop in replacement mostly working out of the box

* Added support for saving and restoring timers

nodes are checked for running timers and timeout and elapsed times are
saved in the nodes meta data.

* Added support for saving and restoring param2 data

fixes bug with ladder orientation being wrong when registered as lightable
nodes.

* Added exile spesific lightable nodes

Added check for the mod exile_env_sounds to confirm we're in exile game
and load exile specific lightable nodes.  Added needed dependancies to mod.conf

* Added registration of tree_mark node as lightable

also added comments

* Added light sources to match illumination levels

registered light sources to match the levels they had in the
illumination mod.

* Bug fix - timer always set at player position

Found the patch to save and restore the node timers wasn't working
because the players possition was always getting a node timer set
even if there was no active light there.  This corrects the problem.

Co-authored-by: Izzy Blacklock <izzy.blacklock@gmail.com>